Move basic Maps Funtion to extendable fragment
