+++ app/src/main/java/com/example/suberduberuber/Fragments/MapFullFragment.java
@@ -56,6 +56,7 @@ import androidx.fragment.app.Fragment;
 import com.example.suberduberuber.Models.Path;
 import com.example.suberduberuber.Models.DroppedPinPlace;
 import com.example.suberduberuber.R;
+import com.example.suberduberuber.Services.MapsServices;
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.GoogleApiAvailability;
 import com.google.android.gms.common.api.Status;
@@ -90,11 +91,9 @@ import static android.app.Activity.RESULT_CANCELED;
 import static android.app.Activity.RESULT_OK;
 
 
-public class MapFullFragment extends Fragment implements OnMapReadyCallback {
+public class MapFullFragment extends MapsServices {
 
     private static final float DEFAULT_ZOOM = 15;
-    private MapView mMapView;
-    private GoogleMap mMap;
     private static final String MAPVIEW_BUNDLE_KEY = "MapViewBundleKey";
     private static final String TAG = "Auto Complete Log";
 
@@ -107,12 +106,10 @@ public class MapFullFragment extends Fragment implements OnMapReadyCallback {
     LatLng mDefaultLocation = new LatLng(53.2734, -7.77832031);
     int AUTOCOMPLETE_REQUEST_CODE = 1;
 
-    FusedLocationProviderClient fusedLocationProviderClient;
-    public boolean locationPermissionGranted = false;
-    public static final int PERMISSIONS_REQUEST_ENABLE_GPS = 9002;
-    public static final int PERMISSIONS_REQUEST_ACCESS_FINE_LOCATION = 9003;
-    public static final int ERROR_DIALOG_REQUEST = 9001;
+
     public static final String ERROR_TAG = "Login Activity";
+
+    private MapsServices mapsServices;
 
     public MapFullFragment() {
         // Required Empty Constructor
@@ -123,13 +120,11 @@ public class MapFullFragment extends Fragment implements OnMapReadyCallback {
         super.onCreateView(inflater, container, savedInstanceState);
         View view = inflater.inflate(R.layout.fragment_map_full, container, false);
 
-        mMapView = (MapView) view.findViewById(R.id.full_map);
         confirmButton = view.findViewById(R.id.confirmButton);
         textView = view.findViewById(R.id.autocomplete);
 
         Places.initialize(getActivity().getApplicationContext(), getString(R.string.google_map_api_key));
         placesClient = Places.createClient(getContext());
-        fusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(getContext());
 
         textView.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -143,12 +138,10 @@ public class MapFullFragment extends Fragment implements OnMapReadyCallback {
                 startActivityForResult(intent, AUTOCOMPLETE_REQUEST_CODE);
             }
         });
-
-        initGoogleMap(savedInstanceState);
         return view;
     }
 
-    private  void updateMarker() {
+    private void updateMarker() {
         if (currentMarker != null) {
             currentMarker.remove();
         }
@@ -160,7 +153,6 @@ public class MapFullFragment extends Fragment implements OnMapReadyCallback {
 
     @Override
     public void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
-        // Get Autocomplete Results
         if (requestCode == AUTOCOMPLETE_REQUEST_CODE) {
             if (resultCode == RESULT_OK) {
                 currentPlace = Autocomplete.getPlaceFromIntent(data);
@@ -175,33 +167,12 @@ public class MapFullFragment extends Fragment implements OnMapReadyCallback {
             } else if (resultCode == RESULT_CANCELED) {
             }
         }
-        // Check For Location Permission
-        else {
-            super.onActivityResult(requestCode, resultCode, data);
-            Log.d(ERROR_TAG, "onActivityResult: called.");
-            switch (requestCode) {
-                case PERMISSIONS_REQUEST_ENABLE_GPS: {
-                    if (locationPermissionGranted) {
-                    } else {
-                        getLocationPermission();
-                    }
-                }
-            }
-        }
+        super.onActivityResult(requestCode, resultCode, data);
     }
 
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-    }
-
-    private void initGoogleMap(Bundle savedInstanceState) {
-        Bundle mapViewBundle = null;
-        if (savedInstanceState != null) {
-            mapViewBundle = savedInstanceState.getBundle(MAPVIEW_BUNDLE_KEY);
-        }
-        mMapView.onCreate(mapViewBundle);
-        mMapView.getMapAsync(this);
     }
 
     @Override
@@ -216,55 +187,33 @@ public class MapFullFragment extends Fragment implements OnMapReadyCallback {
     }
 
     private void getDeviceLocation() {
-
         if (ContextCompat.checkSelfPermission(getActivity().getApplicationContext(), ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
             Class<?> currentClass = this.getClass();
             fusedLocationProviderClient.getLastLocation().addOnCompleteListener(new OnCompleteListener<Location>() {
                 @Override
                 public void onComplete(@NonNull Task<Location> task) {
-                    if(task.isSuccessful()) {
+                    if (task.isSuccessful()) {
                         Location location = task.getResult();
                         LatLng deviceLatLng = new LatLng(location.getLatitude(), location.getLongitude());
                         if (currentClass == SelectOriginFragment.class) {
                             try {
                                 currentPlace = new DroppedPinPlace(deviceLatLng, getContext(), "Current Location").getDroppedPinPlace();
                                 updateMarker();
-                            } catch (IOError | IOException exception) { }
-                        }
-                        else {
+                            } catch (IOError | IOException exception) {
+                            }
+                        } else {
                             mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(deviceLatLng, DEFAULT_ZOOM));
                         }
                     }
                 }
             });
-
-
-            /*
-            SimpleLocation location = new SimpleLocation(getContext());
-            double latitude = location.getLatitude();
-            double longitude = location.getLongitude();
-            LatLng deviceLatLng = new LatLng(latitude, longitude);
-            if (this.getClass() == SelectOriginFragment.class) {
-                try {
-                    currentPlace = new DroppedPinPlace(deviceLatLng, getContext(), "Current Location").getDroppedPinPlace();
-                    updateMarker();
-                } catch (IOError | IOException exception) {
-                    Log.i(TAG, exception.getLocalizedMessage());
-                }
-            }
-            else {
-                mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(deviceLatLng, DEFAULT_ZOOM));
-            }
-             */
         } else {
-            getLocationPermission();
+            permissionsService.getLocationPermission();
         }
     }
 
     @Override
-    public void onMapReady(GoogleMap map) {
-        mMap = map;
-        mMap.setMapType(GoogleMap.MAP_TYPE_NORMAL);
+    public void setupMap() {
         mMap.setOnMapLongClickListener(new GoogleMap.OnMapLongClickListener() {
             @Override
             public void onMapLongClick(LatLng latLng) {
@@ -276,141 +225,15 @@ public class MapFullFragment extends Fragment implements OnMapReadyCallback {
                     } else {
                         Toast.makeText(getContext(), "Invalid Pin Location", Toast.LENGTH_LONG).show();
                     }
-                }
-                catch (IOError | java.io.IOException exception) {
+                } catch (IOError | java.io.IOException exception) {
                     Log.i(TAG, exception.getLocalizedMessage());
                 }
             }
         });
-        map.setMyLocationEnabled(true);
-        map.getUiSettings().setZoomControlsEnabled(true);
+        mMap.setMyLocationEnabled(true);
+        mMap.getUiSettings().setZoomControlsEnabled(true);
         getDeviceLocation();
     }
-    @Override
-    public void onPause() {
-        mMapView.onPause();
-        super.onPause();
-    }
-
-    @Override
-    public void onDestroy() {
-        mMapView.onDestroy();
-        super.onDestroy();
-    }
-
-    @Override
-    public void onLowMemory() {
-        super.onLowMemory();
-        mMapView.onLowMemory();
-    }
-    @Override
-    public void onResume() {
-        super.onResume();
-        mMapView.onResume();
-        if(checkMapServices()) {
-            if(!locationPermissionGranted) {
-                getLocationPermission();
-            }
-        }
-    }
-
-    @Override
-    public void onStart() {
-        super.onStart();
-        mMapView.onStart();
-    }
-
-    @Override
-    public void onStop() {
-        super.onStop();
-        mMapView.onStop();
-    }
-
-    protected void displayPath(Path path) {
-
-    }
-
-    // ALL METHODS BELOW ARE FROM GITHUB
-    // USER: mitchtabian
-    // URL: https://gist.github.com/mitchtabian/2b9a3dffbfdc565b81f8d26b25d059bf
-    // Code is to ask user for permission to use location services for maps
-    private boolean checkMapServices(){
-        if(isServicesOK()){
-            if(isMapsEnabled()){
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private void buildAlertMessageNoGps() {
-        final AlertDialog.Builder builder = new AlertDialog.Builder(getContext());
-        builder.setMessage("This application requires GPS to work properly, do you want to enable it?")
-                .setCancelable(false)
-                .setPositiveButton("Yes", new DialogInterface.OnClickListener() {
-                    public void onClick(@SuppressWarnings("unused") final DialogInterface dialog, @SuppressWarnings("unused") final int id) {
-                        Intent enableGpsIntent = new Intent(android.provider.Settings.ACTION_LOCATION_SOURCE_SETTINGS);
-                        startActivityForResult(enableGpsIntent, PERMISSIONS_REQUEST_ENABLE_GPS);
-                    }
-                });
-        final AlertDialog alert = builder.create();
-        alert.show();
-    }
-
-    public boolean isMapsEnabled(){
-        final LocationManager manager = (LocationManager) getActivity().getSystemService( Context.LOCATION_SERVICE );
-
-        if ( !manager.isProviderEnabled( LocationManager.GPS_PROVIDER ) ) {
-            buildAlertMessageNoGps();
-            return false;
-        }
-        return true;
-    }
-
-    private void getLocationPermission() {
-        if (ContextCompat.checkSelfPermission(getActivity().getApplicationContext(),
-                android.Manifest.permission.ACCESS_FINE_LOCATION)
-                == PackageManager.PERMISSION_GRANTED) {
-            locationPermissionGranted = true;
-        } else {
-            ActivityCompat.requestPermissions(getActivity(),
-                    new String[]{android.Manifest.permission.ACCESS_FINE_LOCATION},
-                    PERMISSIONS_REQUEST_ACCESS_FINE_LOCATION);
-        }
-    }
-
-    public boolean isServicesOK(){
-        Log.d(ERROR_TAG, "isServicesOK: checking google services version");
-
-        int available = GoogleApiAvailability.getInstance().isGooglePlayServicesAvailable(getActivity());
-
-        if(available == ConnectionResult.SUCCESS){
-            Log.d(ERROR_TAG, "isServicesOK: Google Play Services is working");
-            return true;
-        }
-        else if(GoogleApiAvailability.getInstance().isUserResolvableError(available)){
-            Log.d(ERROR_TAG, "isServicesOK: an error occured but we can fix it");
-            Dialog dialog = GoogleApiAvailability.getInstance().getErrorDialog(getActivity(), available, ERROR_DIALOG_REQUEST);
-            dialog.show();
-        }else{
-            Toast.makeText(getActivity(), "You can't make map requests", Toast.LENGTH_SHORT).show();
-        }
-        return false;
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode,
-                                           @NonNull String permissions[],
-                                           @NonNull int[] grantResults) {
-        locationPermissionGranted = false;
-        switch (requestCode) {
-            case PERMISSIONS_REQUEST_ACCESS_FINE_LOCATION: {
-                if (grantResults.length > 0
-                        && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    locationPermissionGranted = true;
-                }
-            }
-        }
-    }
-    // ENDS METHODS FROM GIT HUB FOR LOCATION SERVICES
 }
+
+
