+++ app/src/main/java/com/example/suberduberuber/Services/MapsServices.java
@@ -1,0 +1,124 @@
+package com.example.suberduberuber.Services;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import com.example.suberduberuber.R;
+import com.google.android.gms.location.FusedLocationProviderClient;
+import com.google.android.gms.location.LocationServices;
+import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.MapView;
+import com.google.android.gms.maps.OnMapReadyCallback;
+
+
+public abstract class MapsServices extends Fragment implements OnMapReadyCallback {
+
+    protected MapView mMapView;
+    protected GoogleMap mMap;
+    protected FusedLocationProviderClient fusedLocationProviderClient;
+    protected PermissionsService permissionsService;
+    public boolean locationPermissionGranted = false;
+
+
+    private String TAG = "Maps Services";
+    private static final String MAPVIEW_BUNDLE_KEY = "MapViewBundleKey";
+    public static final int PERMISSIONS_REQUEST_ENABLE_GPS = 9002;
+
+    public MapsServices() {
+
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        super.onCreateView(inflater, container, savedInstanceState);
+        View view = inflater.inflate(R.layout.fragment_map_full, container, false);
+        mMapView = (MapView) view.findViewById(R.id.full_map);
+        fusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(getContext());
+        permissionsService = new PermissionsService(getContext(), getActivity());
+        initMap(savedInstanceState);
+        return view;
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
+        Log.d(TAG, "onActivityResult: called.");
+        switch (requestCode) {
+            case PERMISSIONS_REQUEST_ENABLE_GPS: {
+                if (locationPermissionGranted) {
+                } else {
+                    permissionsService.getLocationPermission();
+                }
+            }
+        }
+    }
+
+    public void initMap(Bundle savedInstanceState) {
+        Bundle mapViewBundle = null;
+        if (savedInstanceState != null) {
+            mapViewBundle = savedInstanceState.getBundle(MAPVIEW_BUNDLE_KEY);
+        }
+        mMapView.onCreate(mapViewBundle);
+        mMapView.getMapAsync(this);
+    }
+
+    public void setupMap() {
+        mMap.setMyLocationEnabled(true);
+        mMap.getUiSettings().setZoomControlsEnabled(true);
+    }
+
+    @Override
+    public void onMapReady(GoogleMap map) {
+        mMap = map;
+        mMap.setMapType(GoogleMap.MAP_TYPE_NORMAL);
+        setupMap();
+    }
+    @Override
+    public void onPause() {
+        mMapView.onPause();
+        super.onPause();
+    }
+
+    @Override
+    public void onDestroy() {
+        mMapView.onDestroy();
+        super.onDestroy();
+    }
+
+    @Override
+    public void onLowMemory() {
+        super.onLowMemory();
+        mMapView.onLowMemory();
+    }
+    @Override
+    public void onResume() {
+        super.onResume();
+        mMapView.onResume();
+        if(permissionsService.checkMapServices()) {
+            if(!locationPermissionGranted) {
+                permissionsService.getLocationPermission();
+            }
+        }
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        mMapView.onStart();
+    }
+
+    @Override
+    public void onStop() {
+        super.onStop();
+        mMapView.onStop();
+    }
+}
