+++ app/src/main/res/navigation/rider_nav_graph.xml
@@ -9,12 +9,7 @@
         android:id="@+id/confirmRideFragment2"
         android:name="com.example.suberduberuber.Fragments.ConfirmRideFragment"
         android:label="fragment_confirm_ride"
-        tools:layout="@layout/fragment_confirm_ride" >
-        <action
-            android:id="@+id/action_confirmRideFragment2_to_ridePendingFragment"
-            app:destination="@id/ridePendingFragment"
-            app:enterAnim="@android:anim/slide_out_right" />
-    </fragment>
+        tools:layout="@layout/fragment_confirm_ride" />
 
     <fragment
         android:id="@+id/rateDriverFragment"
